requirements.txt
streamlit
google-generativeai
python-dotenv
pytest
pytest-mock